style: update CSS in App.css to improve scrollbar appearance feat: add TableComponent to CsvOutputComponent for better table rendering refactor: remove unused code and improve readability in CsvOutputComponent refactor: simplify logic in TableAutoCellRender component feat: add autoHeight property to columns in extractColumnsFromRows utility function
