+++ src/frontend/src/components/tableAutoCellRender/index.tsx
@@ -9,28 +9,36 @@ import StringReader from "../stringReaderComponent";
 export default function TableAutoCellRender({
   value,
 }: CustomCellRendererProps) {
-  switch (typeof value) {
-    case "object":
-      if (value === null) {
+  function getCellType() {
+    switch (typeof value) {
+      case "object":
+        if (value === null) {
+          return String(value);
+        } else if (Array.isArray(value)) {
+          return <ArrayReader array={value} />;
+        } else if (value.definitions) {
+          // use a custom render defined by the sender
+        } else {
+          return <ObjectRender object={value} />;
+        }
+        break;
+      case "string":
+        if (isTimeStampString(value)) {
+          return <DateReader date={value} />;
+        } else {
+          return <StringReader string={value} />;
+        }
+        break;
+      case "number":
+        return <NumberReader number={value} />;
+      default:
         return String(value);
-      } else if (Array.isArray(value)) {
-        return <ArrayReader array={value} />;
-      } else if (value.definitions) {
-        // use a custom render defined by the sender
-      } else {
-        return <ObjectRender object={value} />;
-      }
-      break;
-    case "string":
-      if (isTimeStampString(value)) {
-        return <DateReader date={value} />;
-      } else {
-        return <StringReader string={value} />;
-      }
-      break;
-    case "number":
-      return <NumberReader number={value} />;
-    default:
-      return String(value);
+    }
   }
+
+  return (
+    <div className="group flex h-full w-full items-center align-middle">
+      {getCellType()}
+    </div>
+  );
 }
